implements a multi-level nested netDB with callers for netDB segmentation
This patch adds an interface which is similar to the existing netDB, which is designed to simply 'juggle' multiple
self-contained netDBs internally, so that storage and retrieval can be controlled by passing an 'id' which determines
which netDB to perform an operation on. This allows the caller to provide a contextual identity wherein the operations
are applied, making it impossible to for instance associate a store which was sent to our router with a store which was
sent to our client if they use different contextual identifiers.