Router: Expansive changes to netDb handling which enable the use of the 'Safe' interface which automatically isolates each operation based on where it came from. At this time, it does isolate operations successfully but only in a partial manner. Also, I am fairly sure that at this time, Encrypted leaseSets are broken in this branch. Much of this issue is visible in the Key-Manager functions of the router console(apps/routerconsole, grep for netDb). Family keys might be broken but they probably are not. Isolated netDb stores are reflected in the UI but there are some obvious issues with formatting I could probably ask for help with soon. I don't think we need a separate netDb for the exploratory pool, in fact I think that might be a contradiction, but maybe if it's a 'triage' area? Like some kind of whanau-inspired thing to make comparisons and guess about the integrity of the other DB? Probably a silly idea, but the long and short of it is that the null-client/exploratory netDB is unused.
THE MOST IMPORTANT ISSUE BY FAR though is that for now, when the dbid is null, in some cases, we will search all the clients first, then the floodfill for the data we want, which breaks isolation(albeit predictably). We want to make sure that dbid is null only when we mean it, then change the meaning of null to these functions to always mean floodfill. This will give us 'True isolation'.