Transport: Rework unbanning of peers when reachable.
Currently, when a peer is marked as reachable, it is also
unconditionally unbanned, whether or not the peer was
even banned in the first place.

Since we don't test banned peers for reachability, this
behavior had the consequence of spamming calls to unban
peers who weren't banned to begin with.

In case this section of code has been implemented in unexpected
ways, make adjustments to continue the current behavior,
but highlighted with an warning if we actually run across
cases where HARD banned peers are actually being unbanned.

Signed-off-by: obscuratus <obscuratus@mail.i2p>