NetDb: Local LS checks in subDb context.
Currently, when a LS is stored, we check all the local LS on the
router, and throw an exception if an attempt is made to store a
local LS.

But with segmented netDb, each client netDb has no knowledge of the
other clients, so limit this check to the client netDb, and
proceed to store the LS for other local clients if requested.

Otherwise, local clients won't be able to work with each other.

Signed-off-by: obscuratus <obscuratus@mail.i2p>