NetDb: Update the Handling of DSRM in Segmented NetDb.
It is safe to pass client DSRM back to the inNetMessagePool, even
though they will lose their client context, as long as the replies
have been stripped.

The DSRM will be matched against it's search, and the handling
then ends up back in the client context.  Or the DSRM is dropped.

TODO: The replies can actually be useful.  If the reply RI could
be handled inside the context of the client netDb, with searches
going out through tunnels, they should be safe.

Signed-off-by: obscuratus <obscuratus@mail.i2p>